--- frontend/src/Home.jsx
@@ -1,9 +1,0 @@
-function Home() {
-	return (
-		<div className="HomeScreen">
-		Welcome to unishare, nice to have you here
-		</div>
-	)
-}
-
-export default Home
