+++ frontend/src/pages/Home.jsx
@@ -1,0 +1,9 @@
+function Home() {
+	return (
+		<div className="HomeScreen">
+		Welcome to unishare, nice to have you here
+		</div>
+	)
+}
+
+export default Home
