+++ frontend/src/pages/Registration.jsx
