+++ frontend/src/pages/Tos.jsx
